Small mods to getAccessToken in parkrunAPI.php
When the getAccessToken function retrieves the client_id and client_secret from the .parkrunapi.keys file and assigns to $user and $secret variables superfluous trailing spaces are included. The use of trim as the variables are assigned removes these spaces. 

In the same function, the regular expression that extracted the client_secret from the .parkrunapi.keys file required a newline to be present after the client_secret value. I've simplified this regular expression such that the newline is not required.